Rename FetchTodayOrderTotals to FetchStatsForComplications
