+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/complications/ordertotals/OrderTotalsComplicationService.kt
@@ -6,14 +6,14 @@ import androidx.wear.watchface.complications.datasource.ComplicationRequest
 import androidx.wear.watchface.complications.datasource.SuspendingComplicationDataSourceService
 import com.woocommerce.android.R
 import com.woocommerce.android.wear.complications.createTextComplicationData
-import com.woocommerce.android.wear.complications.ordertotals.FetchTodayOrderTotals.StatType.ORDER_TOTALS
+import com.woocommerce.android.wear.complications.ordertotals.FetchStatsForComplications.StatType.ORDER_TOTALS
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
 class OrderTotalsComplicationService : SuspendingComplicationDataSourceService() {
 
-    @Inject lateinit var fetchTodayOrderTotals: FetchTodayOrderTotals
+    @Inject lateinit var fetchStatsForComplications: FetchStatsForComplications
 
     override fun getPreviewData(type: ComplicationType): ComplicationData? {
         return when (type) {
@@ -31,7 +31,7 @@ class OrderTotalsComplicationService : SuspendingComplicationDataSourceService()
         return when (request.complicationType) {
             ComplicationType.SHORT_TEXT -> createTextComplicationData(
                 context = applicationContext,
-                content = fetchTodayOrderTotals(ORDER_TOTALS),
+                content = fetchStatsForComplications(ORDER_TOTALS),
                 description = getString(R.string.order_totals_complication_preview_description)
             )
